numademo: fix error on 32bit system
Error Info on 32bit:
root@intel-x86:~# numademo -t -e 1M
Configured Nodes does not match available memory nodes

That's because sizeof(long)=4Word, but sizeof(long long)=8Word
on 32bit. So (long long) assigning to (long) maybe cause overflow.

long numa_node_size(int node, long *freep)
{
    ...
    long sz = numa_node_size64_int(node, &f2);
    return sz;
    ...
}
long long numa_node_size64(int node, long long *freep)
{
    ...
}

+++ libnuma.c
@@ -793,10 +793,10 @@ long long numa_node_size64(int node, long long *freep)
 
 make_internal_alias(numa_node_size64);
 
-long numa_node_size(int node, long *freep)
+long long numa_node_size(int node, long long *freep)
 {
 	long long f2;
-	long sz = numa_node_size64_int(node, &f2);
+	long long sz = numa_node_size64_int(node, &f2);
 	if (freep)
 		*freep = f2;
 	return sz;
