+++ numademo.c
@@ -301,7 +301,7 @@ int max_node, numnodes;
 int get_node_list(void)
 {
         int a, got_nodes = 0;
-        long free_node_sizes;
+        long long free_node_sizes;
 
         numnodes = numa_num_configured_nodes();
         node_to_use = (int *)malloc(numnodes * sizeof(int));
